ParseSoapExceptionTest lagt til, rydding gjort
