+++ Difi.Oppslagstjeneste.Klient.Domene/Abstract/EndringsInfo.cs
@@ -2,7 +2,7 @@
 using System.Xml;
 using Difi.Oppslagstjeneste.Klient.Felles.Envelope;
 
-namespace Difi.Oppslagstjeneste.Klient.Domene
+namespace Difi.Oppslagstjeneste.Klient.Domene.Abstract
 {
     public abstract class EndringsInfo
     {
