+++ Difi.Oppslagstjeneste.Klient.Domene/Difi.Oppslagstjeneste.Klient.Domene.csproj
@@ -48,7 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EndringsInfo.cs" />
+    <Compile Include="Abstract\EndringsInfo.cs" />
     <Compile Include="Epostadresse.cs" />
     <Compile Include="Exceptions\OppslagstjenesteException.cs" />
     <Compile Include="Exceptions\SoapException.cs" />
