+++ Difi.Oppslagstjeneste.Klient.Domene/Mobiltelefonnummer.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Xml;
+using Difi.Oppslagstjeneste.Klient.Domene.Abstract;
 
 namespace Difi.Oppslagstjeneste.Klient.Domene
 {
