+++ Difi.Oppslagstjeneste.Klient.Tester/Difi.Oppslagstjeneste.Klient.Tester.csproj
@@ -55,13 +55,11 @@
   <ItemGroup>
     <Compile Include="OppslagstjenesteKonfigurasjonTester.cs" />
     <Compile Include="Svar\EndringerSvarTester.cs" />
-    <Compile Include="Domene\EndringsInfoTester.cs" />
     <Compile Include="Svar\PersonerSvarTester.cs" />
     <Compile Include="Domene\PersonTester.cs" />
     <Compile Include="Svar\PrintSertifikatSvarTester.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Integrasjon\Integrasjonstester.cs" />
-    <Compile Include="Domene\SikkerDigitalPostAdresseTester.cs" />
     <Compile Include="Domene\SoapExceptionTester.cs" />
   </ItemGroup>
   <ItemGroup>
