--- Difi.Oppslagstjeneste.Klient.Tester/Domene/SikkerDigitalPostAdresseTester.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Difi.Oppslagstjeneste.Klient.Tester.Domene
-{
-    [TestClass]
-    public class SikkerDigitalPostAdresseTester
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
