+++ Difi.Oppslagstjeneste.Klient.Tester/OppslagstjenesteKonfigurasjonTester.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Difi.Oppslagstjeneste.Klient.Tester
+{
+    [TestClass]
+    public class OppslagstjenesteKonfigurasjonTester
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
